tdf#130857 qt weld: Set button image set in .ui file
Implement logic to set the image for a button, as specified
via the "image" property for the "GtkButton" object in
the .ui file.

Similar to how VclBuilder::makeObject does it, extract
an icon name and load an image using that one for
"GtkImage" obejcts.
Add another static `toQPixmap` variant that
takes an Image parameter and use that one for conversion.

For buttons, if the "image" property is set, get
the corresponding QLabel object, get the pixmap
from that one and set an icon in the button from that.

With this in place, when opening the "Help" -> "About LibreOfficeDev"
dialog in a WIP branch adding "cui/ui/aboutdialog.ui" to the list
of .ui files in QtInstanceBuilder::IsUIFileSupported, the button
to copy the version information now shows the corresponding icon
when using the qt6 VCL plugin.

Change-Id: If87866d7ab935cbc4162fb513074eefda22c981a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175761
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>