tdf#130857 cui: Move checkbox out of button box
In the "You are switching to the overwrite mode" dialog
that gets shown when pressing the Insert key in Writer,
move the "Do not show again" checkbox above the button
box with the "No" and "Yes" buttons instead of having
it inside of the button box.

This aligns this with how it's done elsewhere, e.g. in
the "LibreOfficeDev Help Not Installed" dialog
(sfx2/uiconfig/ui/helpmanual.ui) that gets shown
when pressing F1 without local help installed.

It also prevents an issue with this case of a checkbox
with quite some text inside a button box resulting in
overlapping UI elements for the native Qt version
of the dialog (once this gets enabled in
QtInstanceBuilder::IsUIFileSupported).

Change-Id: I005da330cca5a8d6ee468477472e0382c288b022
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175791
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>