tdf#130857 qt weld: Support hiding plain text in entry
Evaluate the "visibility" GtkEntry property [1] [2]
and set the QLineEdit's echoMode [3] to QLineEdit::Password
if the visibility is set to false, so that replacement
characters are shown instead of the actual text as
expected.

This is used e.g. in password dialogs, e.g. in Writer:
"File" -> "Properties" -> "Security", press
"Protect..." button.
(But this dialog is not using native Qt widgets yet,
more needs to be implemented first.)

[1] https://docs.gtk.org/gtk3/property.Entry.visibility.html
[2] https://docs.gtk.org/gtk3/method.Entry.set_visibility.html
[3] https://doc.qt.io/qt-6/qlineedit.html#echoMode-prop

Change-Id: I218e4721bbc21f3d6a3553647e92c1ea7cdc4a0f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175806
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>