tdf#162666 Make SfxStateCache handle Items correctly
SfxStateCache had problems processing the extra Item
states disabled/invalid. Note that due to not using a
Pool here it is not possible to use an ItemSet or an
ItemHolder, those would automatically handle these
cases correctly.
In this cache, this currently needs to be done
handish. Also note that because of that this may break
again when changes in the Item/ItemSet/ItemHolder
mechanism may be done.

Change-Id: I007bc25c727e6432062fa0d2af8215a912cb2fff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175871
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>
Tested-by: Jenkins
(cherry picked from commit 288d720e4940d8fdd71715e6d339765e90716931)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175904
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>