call PathSettings::initialize after utl::Bootstrap::reloadData
and add setting 'BRAND_BASE_DIR' alongside existing 'UserInstallation'
which necessitates moving this block after aAppPath is determined

Done so that PathSettings will use the new locations set in bootstrap
for UserInstallation directory and not the locations detected during
preinit.

So autotext can be found relative to the final UserInstallation dir.

Change-Id: Ib6dadbec85a380475537027a8ce3972e31089711
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175960
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176905
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>