tdf#161349 writerfilter: only copy page background vml fill
This fixes my 24.8 regression.

I had this clause in earlier versions of the patchset,
but it moved along with most of the other logic
to the initialization of the default page style.

However, I certainly need to limit copying the properties
of the default page style to ONLY the fill properties here as well.

Change-Id: I168f13ebeea3720eba53703210314f069fb87ae2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/175963
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins