lok: slideshow: send fill, line, font colors for an animated layer
We improved the information collected when a layer to be animated is renderered.
When the fill style is solid we send the fill color for the animated shape, so
that on the client we can check per each texel if it is part of the shape fill
and in case substitute the original fill color with the color computed by the
animation. The same is done for the line color.
Moreover, in case the animated object is a paragraph, we send the font color.

Change-Id: Iffca357879ef13489c5d943fb5c6faeeaca0d1b2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183254
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>