tdf#130857 qt weld: Implement QtInstanceContainer::move
As documented in include/vcl/weld.hxx, the new parent
can be null, in which case the widget is only removed
from the current container. In that case, mark the
widget for deletion, as there's no more parent
that owns the widget and would take care of deleting
it when getting deleted itself.

Change-Id: I1b3374683629a841f5101098b967cc5cf75d4f69
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176090
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>