switching PageKind without switching EditMode retains current page
i.e. ChangeEditMode doesn't do anything if the EditMode is unchanged
but the PageKind was switched from Notes to Standard. So the same
page of the other mode is still selected. An explicit SwitchPage
is required to change the page from the other mode.

This results in the reported mode of "2", but rendering actually still
rendering as if in "mode 0", so a request to render a 'standard' page
returns an image of the currently still-selected 'notes' page, which
confusingly can be cached and not shown until served as a reply to a
later request of the 'standard' page.

See DrawViewShell::ReadFrameViewData and other places for more
of this pattern.

Change-Id: Ie8aa8f93f45189fd6f9c37c4077fa2b547ca4815
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176084
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Gökay ŞATIR <gokaysatir@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
(cherry picked from commit 95e9c210ef8380b0909c6ba596e3023bafef4083)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176824
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins