tdf#130857 qt weld: Add layout/widget for extra msg dialog controls
weld::MessageDialog::weld_message_area needs to return
a weld::Container that can be used to insert additional
controls into a message dialog.

In order to implement this for QtInstanceMessageDialog,
insert an additional widget with a QVBoxLayout layout
into the QMessageBox layout, and return that one.

In order to ensure that the extra controls are inserted
in between the labels holding the text in the message dialog
and the button box, make use of the known implementation detail
that Qt's QMessageBox uses QGridLayout for its layout
(see QMessageBoxPrivate::setupLayout in qtbase [1]),
find where the last label is positioned, shift everything
after the last label down by one row in the grid layout,
then insert the extra widget in the now empty row.

If QMessageBox internals ever change to use a different layout,...,
then this will have to be adjusted in the future, but it
works fine in my tests with both, Qt 5.15 and current
qtbase dev (as of commit 4c0b45553862d3eff35906d02ea5e2afd9252bbd).

This will be needed e.g. by the "LibreOfficDev Help Not Installed"
dialog that shows up when pressing F1 in a build not including
the local help.

[1] https://code.qt.io/cgit/qt/qtbase.git/tree/src/widgets/dialogs/qmessagebox.cpp?id=4c0b45553862d3eff35906d02ea5e2afd9252bbd#n290

Change-Id: I2153add5145a4655800ca89f8e943610014b1021
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176091
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins