slideshow: create a placeholder even if text field content is hidden
A placeholder always needs to be exported even if the content is hidden
since it
could be visible on another slide and master page layers should be
cached on the
client.
In order to support text fields different from the default ones a hash
is used
for linking text field placeholder and text field content in place of
the text
field type.
Any text field content should be sent first, so that when the client try
to draw
a master page layer corresponding to a text field placeholder the
content is
already available.
Slideshow unit tests have been updated to match the new implementation.

Change-Id: I845888d141e0f02c2066a9cb17d89ab152bf28b8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176168
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>