tdf#163764 Force pending timers to run after marked text changes
During native dictation, waiting for the next native event is
blocked while dictation runs in a loop within a native callback.

Because of this, LibreOffice's painting timers won't fire until
dictation is cancelled or the user pauses speaking. So, force
any pending timers to fire after the marked text changes.

Also, remove the fix for OpenOffice bug 106901 as causes any
key down events to cancel dictation and removing the fix does
not appear to cause the original crashing bug to reoccur.

Note: inserting a character from the system Character Viewer
window causes dictation to stop responding and the only way
I have found to restart dictation is by toggling dictation off
and then back on again. This same behavior occurs in Apple's
TextEdit application so this appears to be a macOS bug.

Change-Id: Iad2b54870ff1a315f2f71d72bef24af3cea808e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176100
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>