tdf#163748: checking Any's value type is anti-pattern
The type could be XInterface - yet it may hold XGraphic. Testing
for exact type would fail then.

And it looks like Java bridge code does just that: a code like

  XGraphic graphic = xGraphicProvider.queryGraphic(v);
  xStyleProps.setPropertyValue("BackGraphic", graphic);

sends the XGraphic as an Any with XInterface type (unlike a similar
code in Basic). It's sad that we *move* tests to Python, instead of
having both - it's not like "Python works -> Java would work, too".

Change-Id: I5cd9b1f750e71895232b2c31174ef969a7b15b7c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176149
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>