tdf#130857 qt weld: Support "GtkSeparator"
Handle "GtkSeparator" objects in .ui files:
Create a QFrame and, depending on the orientation
set a shape of either QFrame::HLine ("QFrame draws a
horizontal line that frames nothing (useful as separator)")
or QFrame::VLine ("QFrame draws a vertical line that frames
nothing (useful as separator)") [1].

This will be used e.g. in the "Help" -> "Show Tip of the Day"
dialog.

[1] https://doc.qt.io/qt-6/qframe.html#Shape-enum

Change-Id: I41e69dd211cbb69cb7b23cc54640cd1fad655efc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176165
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins