tdf#130857 qt weld: Declare support for "Tip of the Day" dialog
Add .ui file for the dialog that can be triggered via
"Help" -> "Show Tip of the Day" to the list of supported
.ui files.

This means that native Qt widgets are used for that dialog
now when LO gets started with environment variable
SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This dialog makes use of the QtInstanceDrawingArea
logic implemented in previous commits.

Change-Id: Id252e38e2eb9265276457ab64d582f71d628e369
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176174
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins