tdf#130857 tipofthedaydialog.ui: Move checkbox out of button box
Move the checkbox in the "Help" > "Show Tip of the Day"
dialog above the button box rather than having it inside
of the button box.

That is consistent with e.g. the "LibreOfficeDev Help not installed"
dialog that gets shown when pressing F1 in a build not including
the local help, and it makes this work with the native Qt version
of the dialog (SAL_VCL_QT_USE_WELDED_WIDGETS=1) which doesn't support
arbitrary widgets inside of the button box.

Change-Id: I79541d986df31325b43a17dbec81466ec8bfdca4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176177
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>