tdf#125969 cui: add in-use area image to bitmap list
This fixes a five year old (non-)easyhack with 3 duplicates.
Note the nice-to-have dependencies in the bug report(s).

When the document has a unique background (area) image,
it is now added to the list of available images
IF/WHEN the user looks in the area tab.
This allows the user to switch back after changing,
or re-use it in other places in the document.

Most of this patch ended up being plumbing to ensure that
this added image is ONLY available to the current document,
because it MUST NOT be saved to the user profile.

This change affects all apps and all types of areas: NICE
  -tested Writer pages, paragraphs, headers, textbox, sidebar(page)
  -tested Calc page style
  -tested Draw page, sidebar(page), textbox

Caveats:
-the bitmap list is NOT updated at the time of document import,
 only when area property inspected.
 (The bug requesting inclusion at the time of import is tdf#100832).

make -srj1 UITest_writer_tests8 \
    UITEST_TEST_NAME=tdf125969.tdf125969.test_tdf125969 \
    SAL_USE_VCLPLUGIN=gen

Change-Id: Ic9fea9b199602c4df1376e781d5df019526473d5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176253
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>