tdf#130857 qt weld: Add initial SpinButton support
Add new class QtInstanceSpinButton that is the
weld::SpinButton implementation using native
widgets. Initially, implement the actual logic
of only some methods, and let the others trigger
an assert for now. (These can be implemented once
needed to support more dialogs.)

Initially, assume that the value is an integer
(which is the case for the "Table" -> "Insert"
-> "Rows" dialog in Writer, for which support
will be declared in an upcoming commit).
However, already use a QDoubleSpinBox
that supports floating point values instead
of QSpinBox (that can only handle integer values)
to prepare for extending the implementation
accordingly later.

In QtBuilder, handle "GtkSpinButton" objects,
create a QDoubleSpinBox for them and evaluate
the "digits" and "adjustment" properties to
set the corresponding values for the
QDoubleSpinBox.

Let QtInstanceBuilder::weld_spin_button
return an instance of the newly added
class.

Change-Id: I0808589a3e6bece749c0ae4541f2419410ea99bb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176248
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>