tdf#130857 qt weld: Support sw "Table" -> "Insert" -> "Rows" dialog
Declare support for this dialog seen in Writer when
taking the following steps:

* start Writer
* insert a table using "Table" -> "Insert Table" -> "Insert"
* "Table" -> "Insert" -> "Rows..."

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and LO gets started with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

This dialog makes use of QtInstanceSpinButton
added in the previous commit.

Change-Id: I0ecdf371278c86ddb8d841b1a95c38c52f53d8d8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176249
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>