tdf#155447 a11y: Add test for accessible IDs expected by Orca
Add a new AccessibilityTools::getAccessibleObjectForId
helper that can be used in a11y tests to identify an object
by its accessible ID.

Add a test that checks that the accessible ID of the
spelling dialog matches what Orca's logic expects
in order to identify it, and that more UI elements
with expected accessible IDs exist.

This is to prevent breaking Orca logic without noticing.
Any changes affecting this test should be discussed with
the Orca maintainer first.

See the comment above the newly added test for more details
and a link to the corresponding Orca commits.

The new test initially only checks that objects with corresponding
IDs exist, could be extended to verify more of their
properties as needed.

Logic for checking availability of a dictionary, needed for
spell check was taken over from the existing tdf155705 test in
sw/qa/extras/accessibility/tdf155705.cxx, whose sample doc
is also used.

Change-Id: I0c5629272a89a4a570e080e62e99b6c105369cf9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176281
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>