tdf#130857 qt weld: Extract setting label props
Move logic to set "GtkLabel" properties on
a QLabel to a new helper method
QtBuilder::setLabelProperties and call it right
after creating the QLabel for a "GtkLabel", and only
for that case.

This also prevents potentially overwriting the
text of a QtHyperlinkLabel widget created for a
"GtkLinkButton" object in a .ui file, as that
class also derives from QLabel, but has specialized
handling for setting the label text to be a hyperlink.

Change-Id: I8bfcc31a503a1c9deec32bcd584e1a86c3dfff19
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176296
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins