tdf#130857 qt weld: Always set non-empty link target
In QtHyperlinkLabel::update, always set a non-empty
string for the link target, as otherwise the text
in the QLabel will not be a clickable hyperlink, but
just plain text. While that seems reasonable in general,
this is problematic as the QtHyperlinkLabel is the widget
used for QtInstanceLinkButton, and a weld::LinkButton
should be clickable even if the link target is empty, as
a custom signal handler can be set that doesn't make use
of th URI anyway.

Therefore, just set a single space character as the URI
if the actual one is empty.

This is relevant e.g. for the "LibreOfficeDev Help Not
Installed" dialog that shows when pressing F1 in a build
not including local help, which will be declared as
supported in an upcoming commit.

Change-Id: I1db4c405793109ae9af811d6dd08b1da37246271
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176297
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins