tdf#130857 qt weld: Return actual msg dialog ret code
If a synchronous message dialog is closed by means other than clicking
a button, return the QDialog::exec response code in
QtInstanceMessageDialog::run, comparable to what already
happens in the async case (s. QtInstanceMessageDialog::dialogFinished)
instead of always returning RET_CLOSE.

This is relevant e.g. for the "LibreOfficeDev Help Not
Installed" dialog that shows when pressing F1 in a build
not including local help, which will be declared as
supported in an upcoming commit. That dialog gets
closed with a response code of RET_YES
when the link button is clicked, see
HelpManualMessage's DownloadClickHdl
(in sfx2/source/appl/sfxhelp.cxx).

Change-Id: Iaa1599dda618e61a5b19264f1984d570a45c6881
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176298
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins