tdf#163812 sw: invalidate "Area" properties so sidebar updates
Changes to the background fill were not being updated
in Writer's sidebar, ever since the introduction in 5.3.

Each of these invalidations can be observed to be necessary
in the corresponding fill type and fill choice listboxes.

It already worked in Draw, and I don't see anywhere
this would be needed in Calc

Change-Id: I42f94ee99036761ff69293bcbb44aeda61e1b321
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176300
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>