tdf#130857 qt weld: Support "LibreOffice Help Not Installed" dlg
Declare support for the "LibreOfficeDev Help Not Installed"
dialog that shows when pressing F1 in a build that doesn't
include the local help.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Change-Id: I25254fd462b867a10e3ed6efc31d9b145466b0f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176299
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>