tdf#103916 tdf#163824 cui tparea: always save changed Fill Lists
This was only working for some EditEng components.
This change makes it work for all uses of Area Fill.

I put this in the destructor because the current implementation
(for SvxAreaTabDialog) SavePalettes() on both OK and CANCEL.

I first tried putting it in DeactivatePage,
but that is not called on a CANCEL,
or on a switch to another AREA type.

[ColorListState doesn't need this treatment
since we don't save to disk here,
plus it needs to be shared between SvxAreaTabDialog pages.]

NOTE: the rename/delete fix for bug 157467 is probably a dependency
because this patch will significantly increase the likelihood
that people will try to manage these lists.

Change-Id: I65a84eb6befe91620fd43f3cec2d3ce4dd7ae92d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176325
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>