tdf#130857 VclBuilder: Split XML parsing + applying tab child props
Split the existing VclBuilder::handleTabChild into two
methods and make the logic not specific to vcl::Window
available to other subclasses as well, to make it
available for reuse by QtBuilder in the future.

In order to do that, move the existing XML parsing logic
from the beginning of VclBuilder::handleTabChild to the base
class into WidgetBuilder::handleTabChild.

Add a new purely virtual method `applyTabChildProperties`
to WidgetBuilder and move the corresponding logic for
vcl::Window to the new VclBuilder override of this method.

Call that method at the end of WidgetBuilder::handleTabChild.

For QtBuilder, just add a dummy implementation that
triggers an assert initially, which matches what
would have happened right at the beginning
of the previous WidgetBuilder::handleTabChild
implementation without this commit in place.

Change-Id: Ie5664bd341182fa51035b547accf9393d65a0702
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176320
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>