tdf#163734 Revert commit f4c2c7c79cfe4464ac596afda37b8904d06969db
Commit f4c2c7c79cfe4464ac596afda37b8904d06969db lowered the Skia
flushing priority to solve tdf#157312. But the commit caused the
"window will dock in rectangle" border to not be drawn when dragging
a dockable window over a dockable position.

Surprisingly, tdf#157312 does not reoccur without the commit so
it appears that the commit is now unnecessary.

Change-Id: Ibd74ef94e8448cf32fd690d7a26dd098191040b8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176341
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
(cherry picked from commit 5a38e4f9798c5ff247aa57581adf2671485627fd)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176346
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>